backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apify_client==1.8.1
apify_shared==1.1.2
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
composio_core==0.5.23
composio_openai==0.5.23
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
e2b==0.17.2a59
e2b-code-interpreter==0.0.10
ecdsa==0.19.0
fastapi==0.114.1
firecrawl-py==1.2.3
frozenlist==1.4.1
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gql==3.5.0
graphql-core==3.2.4
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
inflection==0.5.1
instaloader==4.13.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.117
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
more-itertools==10.5.0
motor==3.5.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.44.1
orjson==3.10.7
packaging==24.1
paramiko==3.5.0
passlib==1.7.4
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.4
pyperclip==1.9.0
Pysher==1.0.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
semver==3.0.2
sentry-sdk==2.14.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.38.5
stripe==10.12.0
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
websocket-client==1.8.0
websockets==13.0.1
yarl==1.11.1
zipp==3.20.2
reportlab
notion-client
celery==5.2.7
redis==4.5.1

